feat(*): Proof of Concept
This is code pulled in from @joshdmiller's other repositories and re-committed here in a much
cleaner fashion, so there is no commit history prior to this point.

This is also pre-alpha. There are almost certainly severe bugs, there are certainly missing
features, and if you use it you are at risk of a potentially fatal ostrich attack, among other
things. Trust me - they're vicious.