feat(flow): Flows are now updateable
Running `warlock.flow( 'flowname' )` multiple times will now update the flow. This allows
plugins/spells to be loaded in any order without worrying about dependencies. A flow that has no
options will not validate, and an appropriate error will be thrown.

An opportunity here is to find a way to guess at the intention of the user and throw an error during
validation if a flow shouldn't exist (e.g. if it has no options whatsoever).