feat(flows): Added `collect` option
`{ collect: true }` will ensure that before the step is processed, all stream data is combined into
a single array and passed all at once to a single call. This will also cause the stream to be
"re-flattened" to its original state so future steps will operate nonethewiser.

This also contains a minor fix for stream merging; the correct merge order is guaranteed by first
collecting all in advance, identically to the above. This has a small performance penalty and may
need to be passed as an option in the future.